+++ UnamDownloader/Resources/Compilers/template.manifest-miner
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8" ?> 
+<asmv1:assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1" xmlns:asmv1="urn:schemas-microsoft-com:asm.v1" xmlns:asmv2="urn:schemas-microsoft-com:asm.v2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <assemblyIdentity version="#MANIFESTVERSION" name="#MANIFESTNAME" />
+    <trustInfo xmlns="urn:schemas-microsoft-com:asm.v2">
+        <security>
+            <requestedPrivileges xmlns="urn:schemas-microsoft-com:asm.v3">
+                <requestedExecutionLevel level="#MANIFESTLEVEL" uiAccess="false" />
+            </requestedPrivileges>
+        </security>
+    </trustInfo>
+</asmv1:assembly>
