--- UnamDownloader/Resources/resource.rc
@@ -1,31 +1,0 @@
-#ifdef DefIcon
-MAINICON ICON "#ICON"
-#endif
-#ifdef DefAdmin
-1 24 "administrator.manifest"
-#endif
-#ifdef DefAssembly
-1 VERSIONINFO
-FILEVERSION     #VERSION
-PRODUCTVERSION  #VERSION
-BEGIN
-    BLOCK "StringFileInfo"
-    BEGIN
-        BLOCK "040904b0"
-        BEGIN
-            VALUE "CompanyName", "#COMPANY"
-			VALUE "FileTitle", "#TITLE"
-            VALUE "FileDescription", "#DESCRIPTION"
-            VALUE "FileVersion", "#VERSION"
-            VALUE "LegalCopyright", "#COPYRIGHT"
-			VALUE "LegalTrademark", "#TRADEMARK"
-            VALUE "ProductName", "#PRODUCT"
-            VALUE "ProductVersion", "#VERSION"
-        END
-    END
-    BLOCK "VarFileInfo"
-    BEGIN
-        VALUE "Translation", 0x409, 1200
-    END
-END
-#endif
